Got everything working again

+++ src/functions/dashboard-keys-functions.js
@@ -3,6 +3,7 @@ import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
 import { DynamoDBDocumentClient, GetCommand, QueryCommand, UpdateCommand, DeleteCommand, PutCommand } from "@aws-sdk/lib-dynamodb";
 import crypto from "crypto";
 import dotenv from "dotenv";
+import { getAuthThenUser } from "./dashboard-login-functions.js";
 dotenv.config();
 
 
@@ -11,29 +12,6 @@ const client = new DynamoDBClient({ region: process.env.AWS_REGION });
 const dynamo = DynamoDBDocumentClient.from(client);
 const KEYS_TABLE = "KeyTable";
 // const USERS_TABLE = "DyceTable";
-
-// Helper function to get a user by email
-// const getUserByEmail = async (email) => {
-//     const command = new QueryCommand({
-//         TableName: USERS_TABLE,
-//         IndexName: "email-index",
-//         KeyConditionExpression: "email = :email",
-//         ExpressionAttributeValues: { ":email": email },
-//     });
-
-//     const { Items } = await dynamo.send(command);
-//     return Items.length ? Items[0] : null;
-// };
-
-// Helper function to get a user by id
-// const getUserById = async (id) => {
-//     const getUserCommand = new GetCommand({
-//         TableName: USERS_TABLE,
-//         Key: { id: id },
-//     });
-//     const { Item } = await dynamo.send(getUserCommand);
-//     return Item;
-// }
 
 // Helper function to get an API key data
 const getKey = async (key) => {
@@ -58,26 +36,15 @@ const getKeysByUser = async (id) => {
     return Items.length ? Items : null;
 }
 
-// Middleware requiring authentication
-// const requireAuth = async (req, res, next) => {
-//     if (!req.session.userId)
-//         return res.status(401).json({ message: "You must be logged in to access this route" });
-//     const user = await getUserById(req.session.userId);
-//     if (!user)
-//         return res.status(404).json({ message: "User not found" });
-//     req.user = user;
-//     next();
-// };
-
 // Generate API key route
 export const generateApiKey = async (event) => {
     console.log("Received generate key request");
-    const { name, user } = JSON.parse(event.body);
-
+    const { name } = JSON.parse(event.body);
     if (!name) return { statusCode: 400, body: JSON.stringify({ message: "Key name required" }) };
-    if (!user) return { statusCode: 400, body: JSON.stringify({ message: "User not found" }) };
+    const user = await getAuthThenUser(event);
+    if (!user) return { statusCode: 401, body: JSON.stringify({ message: "Unauthorized" }) };
 
-    const userKeys = await getKeysByUser(user);
+    const userKeys = await getKeysByUser(user.id);
     if (userKeys?.some(apiKey => apiKey.name === name)) {
         return { statusCode: 400, body: JSON.stringify({ message: "Key name already in use!" }) };
     }
@@ -87,26 +54,23 @@ export const generateApiKey = async (event) => {
         TableName: KEYS_TABLE,
         Item: {
             key: newKey,
-            userId: user,
+            userId: user.id,
             name: name,
             wallet: user.wallets.length ? user.wallets[0].name : null,
             useCount: 0,
         },
     });
     await dynamo.send(command);
-    console.log(newKey);
     return { statusCode: 200, body: JSON.stringify({ apiKey: newKey }) };
 };
 
 // Get API keys route
 export const getApiKeys = async (event) => {
     console.log("Received get keys request");
-    console.log(event)
-    const { user } = JSON.parse(event.body);
-    console.log(user);
-    if (!user) return { statusCode: 400, body: JSON.stringify({ message: "User not found" }) };
+    const user = await getAuthThenUser(event);
+    if (!user) return { statusCode: 401, body: JSON.stringify({ message: "Unauthorized" }) };
 
-    const apiKeys = await getKeysByUser(user);
+    const apiKeys = await getKeysByUser(user.id);
     const formattedKeys = apiKeys?.map(({ key, name, wallet, useCount }) => ({ //userId,
         name,
         key: `${key.slice(0, 4)}...${key.slice(-4)}`,
@@ -119,14 +83,16 @@ export const getApiKeys = async (event) => {
 // Revoke API key route
 export const deleteApiKey = async (event) => {
     console.log("Received delete key request");
+    const user = await getAuthThenUser(event);
+    if (!user) return { statusCode: 401, body: JSON.stringify({ message: "Unauthorized" }) };
 
-    const { name, user } = JSON.parse(event.body);
+    const { name } = JSON.parse(event.body);
     if (!name) return { statusCode: 400, body: JSON.stringify({ message: "API key name required" }) };
     if (!user) return { statusCode: 400, body: JSON.stringify({ message: "User not found" }) };
 
-    const userKeys = await getKeysByUser(user);
+    const userKeys = await getKeysByUser(user.id);
     const apiKey = userKeys.find(apiKey => apiKey.name === name);
-    
+
     const command = new DeleteCommand({
         TableName: KEYS_TABLE,
         Key: { key: apiKey.key },
@@ -136,38 +102,14 @@ export const deleteApiKey = async (event) => {
     return { statusCode: 200, body: JSON.stringify({ message: "API key deleted successfully" }) };
 };
 
-// Middleware requiring API key
-// const requireKey = async (req, res, next) => {
-//     const key = req.headers['apikey'];
-
-//     if (!key)
-//         return res.status(401).json({ message: "API key required" });
-    
-//     const apiKey = await getKey(key);
-//     if (!apiKey)
-//         return res.status(401).json({ message: "Invalid API key" });
-
-//     const command = new UpdateCommand({
-//         TableName: KEYS_TABLE,
-//         Key: { key: apiKey.key },
-//         UpdateExpression: "SET useCount = :useCount",
-//         ExpressionAttributeValues: {
-//             ":useCount": apiKey.useCount + 1
-//         }
-//     })
-//     await dynamo.send(command);
-
-//     req.apiKey = apiKey;
-//     next();
-// };
 
 // Use API key route (example)
 export const useApiKey = async (event) => {
     console.log("Received use key request");
-
-    const key = event.headers.apikey;
+    const { key } = JSON.parse(event.body);
     if (!key) return { statusCode: 401, body: JSON.stringify({ message: "API key required" }) };
-
+    const user = await getAuthThenUser(event);
+    if (!user) return { statusCode: 401, body: JSON.stringify({ message: "Unauthorized" }) };
     const apiKey = await getKey(key);
     if (!apiKey) return { statusCode: 401, body: JSON.stringify({ message: "Invalid API key" }) };
 
@@ -185,13 +127,18 @@ export const useApiKey = async (event) => {
 // Set API key wallet route
 export const setWallet = async (event) => {
     console.log("Received set wallet request");
+    const user = await getAuthThenUser(event);
+    if (!user) return { statusCode: 401, body: JSON.stringify({ message: "Unauthorized" }) };
 
-    const { keyName, walletName, user } = JSON.parse(event.body);
+    const { keyName, walletName } = JSON.parse(event.body);
+    if (!keyName) return { statusCode: 400, body: JSON.stringify({ message: "API key name required" }) };
+    if (!walletName) return { statusCode: 400, body: JSON.stringify({ message: "Wallet name required" }) };
     const wallet = user.wallets.find(wallet => wallet.name === walletName);
     if (!wallet) return { statusCode: 400, body: JSON.stringify({ message: "Wallet not found" }) };
-
     if (!user) return { statusCode: 400, body: JSON.stringify({ message: "User not found" }) };
-    const userKeys = await getKeysByUser(user);
+
+    const userKeys = await getKeysByUser(user.id);
+    if (!userKeys) return { statusCode: 400, body: JSON.stringify({ message: "API keys not found" }) };
     const key = userKeys.find(apiKey => apiKey.name === keyName)?.key;
     if (!key) return { statusCode: 404, body: JSON.stringify({ message: "API key not found" }) };
 
@@ -202,13 +149,14 @@ export const setWallet = async (event) => {
         ExpressionAttributeValues: { ":wallet": wallet.name },
     });
     await dynamo.send(command);
-
     return { statusCode: 200, body: JSON.stringify({ message: "Set wallet successfully" }) };
 };
 
 export const getWallet = async (event) => {
     console.log("Received get wallet address request");
-    const key = event.headers.apikey;
+    const user = await getAuthThenUser(event);
+    if (!user) return { statusCode: 401, body: JSON.stringify({ message: "Unauthorized" }) };
+    const {key} = JSON.parse(event.body);
     if (!key) return { statusCode: 401, body: JSON.stringify({ message: "API key required" }) };
 
     const apiKey = await getKey(key);
